backend/requirements.txt
Django==2.2.28
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.1
django-rest-passwordreset==1.3.0
djongo==1.3.1
pymongo==3.12.3